refactor(sidepanel): separate live search input from debounced filter state (#53)
- introduce debouncedSearchQuery in useSidepanelData, synced from searchQuery via watch + useDebounceFn
- remove conflicting v-model + @update:search-query dual binding in Sidepanel.vue
- expose clearSearch() that resets searchQuery, debouncedSearchQuery and compactMode immediately
- extract shared test helpers (buildSampleTree) into testUtils.ts
- update NIT_ROADMAP: mark test helper deduplication as completed
