+++ src/sidepanel/Sidepanel.vue
@@ -26,7 +26,7 @@ watchEffect(() => {
 })
 
 // --- Logic Composables ---
-const { structuredMarks, searchQuery, compactMode, setSearchQuery, filteredTree, refreshAllMarks } = useSidepanelData()
+const { structuredMarks, searchQuery, compactMode, clearSearch, filteredTree, refreshAllMarks } = useSidepanelData()
 const {
   collapsedStates,
   collapsedUrls,
@@ -134,14 +134,6 @@ function isGroupCollapsed(url: string, groupTitle: string, totalMarks: number): 
   return state !== undefined ? state : totalMarks > 15
 }
 
-function handleSearchInput(value: string) {
-  setSearchQuery(value)
-}
-
-function handleClearSearch() {
-  setSearchQuery('')
-}
-
 function handleOpenOptions() {
   browser.runtime.openOptionsPage()
 }
@@ -206,7 +198,7 @@ async function handleDeleteTag(tagId: string) {
       @open-options="handleOpenOptions"
       @start-creating-tag="startCreatingTag"
       @cancel-creating-tag="cancelCreatingTag"
-      @clear-search="handleClearSearch"
+      @clear-search="clearSearch"
     />
 
     <div
@@ -287,7 +279,7 @@ async function handleDeleteTag(tagId: string) {
           <p>未找到包含「{{ searchQuery }}」的标记</p>
           <button
             class="mt-2 text-sm text-blue-600 hover:text-blue-700 dark:text-blue-400 dark:hover:text-blue-300"
-            @click="handleClearSearch"
+            @click="clearSearch"
           >
             清除搜索
           </button>
